third_party/blink/web_tests/fast/dom/shadow/reinsert-insertion-point.html

<!DOCTYPE html>

<script src="../../../resources/js-test.js"></script>

<div id="inside">Should be visible</div>

<script>
var inside = document.getElementById('inside');
var insideRoot = inside.attachShadow({mode: 'open'});

var slot = insideRoot.appendChild(document.createElement('slot'));

shouldBeTrue("inside.offsetHeight > 0");

slot.remove();
insideRoot.appendChild(slot);

shouldBeTrue("inside.offsetHeight > 0");
</script>
