third_party/blink/web_tests/fast/dom/shadow/reinsert-insertion-point-expected.txt

PASS inside.offsetHeight > 0 is true
PASS inside.offsetHeight > 0 is true
PASS successfullyParsed is true

TEST COMPLETE

Should be visible
